news_portal/requirements.txt
asgiref==3.8.1
black==25.1.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.3
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
flake8==7.3.0
mariadb==1.1.12
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pycodestyle==2.14.0
pyflakes==3.4.0
reportlab==4.4.2
sqlparse==0.5.3
tzdata==2025.2
